requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1.1
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
static3==0.7.0
